[Keyboard] we27: fixbug (arising from suggestions)
