+++ keyboards/wekey/we27/info.json
@@ -5,7 +5,7 @@
     "width": 5, 
     "height": 6.25, 
     "layouts": {
-        "LAYOUT": {
+        "LAYOUT_numpad_6x5": {
             "layout": [
                 {"label":"$", "x":0, "y":0}, {"label":"(", "x":1, "y":0}, {"label":")", "x":2, "y":0}, {"label":"FN", "x":3, "y":0}, {"label":"DEL", "x":4, "y":0}, 
                 {"label":"NLCK", "x":0, "y":1.25}, {"label":"ESC", "x":1, "y":1.25}, {"label":"/", "x":2, "y":1.25}, {"label":"*", "x":3, "y":1.25}, {"label":"-", "x":4, "y":1.25}, 
@@ -15,15 +15,14 @@
                 {"label":"BACK", "x":0, "y":5.25}, {"label":"0", "x":1, "y":5.25, "w":2}, {"label":".", "x":3, "y":5.25}
             ]
         },
-        "LAYOUT_27_encoder": {
+        "LAYOUT_numpad_6x5_encoder": {
             "layout": [
-                {"label":"$", "x":0, "y":0.75}, {"label":"(", "x":1, "y":0.75}, {"label":")", "x":2, "y":0.75}, {"label":"FN", "x":3, "y":0.75},
+                {"label":"$", "x":0, "y":0.75}, {"label":"(", "x":1, "y":0.75}, {"label":")", "x":2, "y":0.75}, {"label":"FN", "x":3, "y":0.75}, {"label":"MUTE", "x":4.75, "y":0.5, "w":0.75, "h":0.75}, {"label":"VOLD", "x":4.25, "y":0, "w":1.25, "h":0.5}, {"label":"VOLU", "x":4.75, "y":1.25, "w":1.25, "h":0.5},
                 {"label":"NLCK", "x":0, "y":2}, {"label":"ESC", "x":1, "y":2}, {"label":"/", "x":2, "y":2}, {"label":"*", "x":3, "y":2}, {"label":"-", "x":4, "y":2}, 
                 {"label":"=", "x":0, "y":3}, {"label":"7", "x":1, "y":3}, {"label":"8", "x":2, "y":3}, {"label":"9", "x":3, "y":3}, {"label":"+", "x":4, "y":3, "h":2}, 
                 {"label":"TAB", "x":0, "y":4}, {"label":"4", "x":1, "y":4}, {"label":"5", "x":2, "y":4}, {"label":"6", "x":3, "y":4}, 
                 {"label":"CAPS", "x":0, "y":5}, {"label":"1", "x":1, "y":5}, {"label":"2", "x":2, "y":5}, {"label":"3", "x":3, "y":5}, {"label":"ENTER", "x":4, "y":5, "h":2}, 
-                {"label":"BACK", "x":0, "y":6}, {"label":"0", "x":1, "y":6, "w":2}, {"label":".", "x":3, "y":6},
-                {"label":"VOLD", "x":4.25, "y":0, "w":1.25, "h":0.5}, {"label":"MUTE", "x":4.75, "y":0.5, "w":0.75, "h":0.75}, {"label":"VOLU", "x":4.75, "y":1.25, "w":1.25, "h":0.5}
+                {"label":"BACK", "x":0, "y":6}, {"label":"0", "x":1, "y":6, "w":2}, {"label":".", "x":3, "y":6}
             ]
         }
     }
