+++ keyboards/wekey/we27/we27.h
@@ -20,13 +20,13 @@
 
 #define ___ KC_NO
 
-#define LAYOUT_numpad_6x5( \
+#define LAYOUT_numpad_6x5(   \
     K00, K01, K02, K03, K04, \
     K10, K11, K12, K13, K14, \
     K20, K21, K22, K23,      \
     K30, K31, K32, K33, K24, \
     K40, K41, K42, K43,      \
-    K50, K51,      K53, K44,\
+    K50, K51,      K53, K44  \
 ) { \
     { K00, K01, K02, K03, K04 }, \
     { K10, K11, K12, K13, K14 }, \
@@ -36,13 +36,13 @@
     { K50, K51, ___, K53, ___ }  \
 }
 
-#define LAYOUT_numpad_6x5_encoder( \
-    K00, K01, K02, K03, K04, K34\
-    K10, K11, K12, K13, K14, K54\
-    K20, K21, K22, K23,      \
-    K30, K31, K32, K33, K24, \
-    K40, K41, K42, K43,      \
-    K50, K51,      K53, K44 \
+#define LAYOUT_numpad_6x5_encoder(      \
+    K00, K01, K02, K03, K04, K34, K54,  \
+    K10, K11, K12, K13, K14,            \
+    K20, K21, K22, K23,                 \
+    K30, K31, K32, K33, K24,            \
+    K40, K41, K42, K43,                 \
+    K50, K51,      K53, K44             \
 ) { \
     { K00, K01, K02, K03, K04 }, \
     { K10, K11, K12, K13, K14 }, \
